Node and doc changes

+++ docker/run_container.sh
@@ -1,32 +1,37 @@
 #!/bin/bash
-
-docker rm -f foundationposev2
-DIR=$(pwd)/
-xhost +local:docker
-docker run \
-    --name foundationposev2 \
-    --gpus all \
-    --env NVIDIA_DISABLE_REQUIRE=1 \
-    -it \
-    --net host \
-    --ipc host \
-    --pid host \
-    --name foundationposev2 \
-    --cap-add=SYS_PTRACE \
-    --security-opt seccomp=unconfined \
-    -v $DIR:$DIR \
-    -v /home:/home \
-    -v /mnt:/mnt \
-    -v /tmp/.X11-unix:/tmp/.X11-unix \
-    -v /tmp:/tmp \
-    -e DISPLAY=${DISPLAY} \
-    -e GIT_INDEX_FILE \
-    -e ROS_DOMAIN_ID=2\
-    -v $(pwd)/configs/:/xml_configs \
-    -e RMW_IMPLEMENTATION=rmw_cyclonedds_cpp\
-    -e CYCLONEDDS_URI=/xml_configs/cyclonedds.xml\
-    foundationposev2:latest \
-    bash -c "cd $DIR && bash"
+IsRunning=`docker ps -f name=foundationposev2 | grep -c "foundationposev2"`;
+if [ $IsRunning -eq "0" ]; then
+    docker rm -f foundationposev2
+    DIR=$(pwd)/
+    xhost +local:docker
+    docker run \
+        --name foundationposev2 \
+        --gpus all \
+        --env NVIDIA_DISABLE_REQUIRE=1 \
+        -it \
+        --net host \
+        --ipc host \
+        --pid host \
+        --name foundationposev2 \
+        --cap-add=SYS_PTRACE \
+        --security-opt seccomp=unconfined \
+        -v $DIR:$DIR \
+        -v /home:/home \
+        -v /mnt:/mnt \
+        -v /tmp/.X11-unix:/tmp/.X11-unix \
+        -v /tmp:/tmp \
+        -e DISPLAY=${DISPLAY} \
+        -e GIT_INDEX_FILE \
+        -e ROS_DOMAIN_ID=2\
+        -v $(pwd)/configs/:/xml_configs \
+        -e RMW_IMPLEMENTATION=rmw_cyclonedds_cpp\
+        -e CYCLONEDDS_URI=/xml_configs/cyclonedds.xml\
+        foundationposev2:latest \
+        bash -c "cd $DIR && bash"
+else
+    echo "Docker image is already running. Opening new terminal...";
+    docker exec -ti foundationposev2 /bin/bash
+fi
 
 # IsRunning=`docker ps -f name=foundationposev2 | grep -c "foundationposev2"`;
 # if [ $IsRunning -eq "0" ]; then
